requirements.txt
python-ffmpeg
youtube-dl